% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepfuns.R
\name{disc_sub}
\alias{disc_sub}
\title{Reduce discipline categories}
\usage{
disc_sub(discs)
}
\arguments{
\item{discs}{Disciplines (vector)}
}
\value{
One discipline category per case
}
\description{
Reduces the discipline category to 1 for easier grouping.
}
