\name{jprep-package}
\alias{jprep-package}
\alias{jprep}
\docType{package}
\title{
\packageTitle{jprep}
}
\description{
\packageDescription{jprep}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{jprep}
\packageIndices{jprep}

}
\author{
\packageAuthor{jprep}

Maintainer: \packageMaintainer{jprep}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{
# simple examples of the most important functions
}
