% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keywords}
\alias{keywords}
\title{Keywords for preprocessing}
\format{
A vector containing keywords
}
\source{
manually made
}
\usage{
keywords
}
\description{
Keywords for preprocessing
}
\keyword{datasets}
