% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepfuns.R
\name{remove_special}
\alias{remove_special}
\title{Remove arbitrary punctuation and special characters}
\usage{
remove_special(x, ...)
}
\arguments{
\item{content}{Article content (character)}

\item{pattern}{A character pattern to remove e.g. Regexp expression}
}
\value{
Article content without the unwanted characters
}
\description{
A universal function for removing arbitrary punctuation and special characters.
}
