% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepfuns.R
\name{unfreq_term_remove}
\alias{unfreq_term_remove}
\title{Remove unfrequent terms}
\usage{
unfreq_term_remove(content, unfreq_terms)
}
\arguments{
\item{content}{Article content (character)}

\item{unfreq_terms}{Unfrequent terms to remove from article content (vector)}
}
\value{
Article content without the unfrequent words
}
\description{
Removes pre-defined unfrequent terms from article content
}
