% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{keywords}
\alias{keywords}
\title{Keywords for preprocessing}
\format{
A vector containing keywords
}
\source{
"C:\Users\soirk\Krisztian\Research\missing_data_paper\R_stuff\jprep\data\"
}
\usage{
keywords
}
\description{
Keywords for preprocessing
}
\keyword{datasets}
